+++ common/src/main/java/com/alien/common/gameplay/entity/living/alien/ovipositor/Ovipositor.java
@@ -1,11 +1,12 @@
 package com.alien.common.gameplay.entity.living.alien.ovipositor;
 
-import com.alien.common.util.AlienHurtUtil;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.Mob;
 import net.minecraft.world.level.Level;
 import org.jetbrains.annotations.NotNull;
+
+import com.avp.common.registry.tag.AVPDamageTypesTags;
 
 public class Ovipositor extends Mob {
 
@@ -20,12 +21,11 @@ public class Ovipositor extends Mob {
 
     // TODO: Replace with a proper tag.
     @Override
-    public boolean hurt(@NotNull DamageSource source, float amount) {
-        if (AlienHurtUtil.isNonDamagingSource(source)) {
-            // Ovipositors should not drown, freeze or suffocate.
+    public boolean hurt(@NotNull DamageSource damageSource, float amount) {
+        if (damageSource.is(AVPDamageTypesTags.DOES_NOT_HURT_ALIENS)) {
             return false;
         }
 
-        return super.hurt(source, amount);
+        return super.hurt(damageSource, amount);
     }
 }